Fix bug with timezone
