+++ requirements.txt
@@ -2,3 +2,4 @@ python-telegram-bot==8.1.1
 requests
 lxml
 datetime
+pytz
